% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive_content}
\alias{archive_content}
\title{archive content for attributes no longer in the model}
\usage{
archive_content(model)
}
\arguments{
\item{model}{a data.frame object containing all data model content.}
}
\description{
Main function for archiving content for attributes no longer included in the data model
}
