% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attribute_myTable}
\alias{attribute_myTable}
\title{attribute_myTable}
\format{
An object of class \code{character} of length 26.
}
\usage{
attribute_myTable
}
\description{
A character vector containing the java and html lines for defining an attribute DataTable
}
\keyword{datasets}
