% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_camel_case}
\alias{get_camel_case}
\title{convert a Attribute string to desired CamelCase syntax with biological abbreviations intact}
\usage{
get_camel_case(x)
}
\arguments{
\item{x}{a string indicating the attribute string}
}
\value{
a string with the attribute name in CamelCase with biological abbreviations intact
}
\description{
convert a Attribute string to desired CamelCase syntax with biological abbreviations intact
}
