% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_title_snake}
\alias{get_title_snake}
\title{convert a Attribute string to desired snake case syntax}
\usage{
get_title_snake(x)
}
\arguments{
\item{x}{a string indicating the attribute string}
}
\value{
a string with the attribute name in snake case
}
\description{
convert a Attribute string to desired snake case syntax
}
