% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_validVals}
\alias{get_validVals}
\title{Return character vector of all valid value strings defined in the data model}
\usage{
get_validVals(df)
}
\arguments{
\item{model}{a data.frame object containing the data model}
}
\value{
a character vector of all valid values defined in the data model
}
\description{
A util function that parses all valid values from a data model
}
