% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JustTheDocDataDictionary-package.R
\docType{package}
\name{JustTheDocsDataDictionary-package}
\alias{JustTheDocsDataDictionary}
\alias{JustTheDocsDataDictionary-package}
\title{JustTheDocsDataDictionary: An R Package To Generate JustTheDocs-themed Schematic Data Model Dictionary Site on GitHub Pages}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Sage-Bionetworks/JustTheDocsDataDictionary}
  \item Report bugs at \url{https://github.com/Sage-Bionetworks/JustTheDocsDataDictionary/issues}
}

}
\author{
\strong{Maintainer}: Jessica Vera \email{jessica.vera@sagebase.org} (\href{https://orcid.org/0000-0001-7411-8501}{ORCID})

}
\keyword{internal}
