% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_contexts.R
\name{main_contexts}
\alias{main_contexts}
\title{Main}
\usage{
main_contexts(repo_raw_url)
}
\arguments{
\item{data_model_url}{a string indicating the https://raw.githubusercontent.com URL of the data model csv to be used.}
}
\description{
A function that executes the whole workflow of creating/updating gh-pages content from a data model.
}
