% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{makeAttributeContent}
\alias{makeAttributeContent}
\title{Make content detailing metadata attributes in the data model}
\usage{
makeAttributeContent(model)
}
\arguments{
\item{model}{a data.frame object containing the data model.}
}
\description{
A function that executes a series of steps to create/update content detailing metadata attributes.
}
