% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectMetadataAttributes}
\alias{selectMetadataAttributes}
\title{Return all rows in model that define a model attribute that isn't a template}
\usage{
selectMetadataAttributes(model)
}
\arguments{
\item{model}{a data.frame object containing the data model}
}
\value{
a subset of `model` that contains all rows that define metadata attributes with `rank` column for ordering attribute md pages on sidebar.
}
\description{
A util function that selects all rows in the data model that define metadata attributes
}
