% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectMetadataTemplates}
\alias{selectMetadataTemplates}
\title{Return all rows in model that define metadata templates}
\usage{
selectMetadataTemplates(model)
}
\arguments{
\item{model}{a data.frame object containing the data model.}
}
\value{
a subset of `model` that contains all rows that define metadata templates.
}
\description{
A util function that selects all rows in the data model that define metadata templates.
}
