% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_model_csv}
\alias{write_model_csv}
\title{write csv in desired format}
\usage{
write_model_csv(df, fid)
}
\arguments{
\item{df}{a data.frame object to be written to csv}

\item{fid}{a string indicating the filename to write to}
}
\description{
write csv in desired format
}
