% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_management.R
\name{get_yaml_header}
\alias{get_yaml_header}
\title{write md yaml header}
\usage{
get_yaml_header(title, parent = NULL, nav_order = NULL)
}
\arguments{
\item{title}{data model attribute string}

\item{parent}{class of markdown file that is being created, options c('Metadata Templates', 'Attributes')}
}
\value{
A multi-line string detailing the markdown yaml header content
}
\description{
write md yaml header
}
