% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Main}
\usage{
main(data_model_url)
}
\arguments{
\item{data_model_url}{a string indicating the https://raw.githubusercontent.com URL of the data model csv to be used.}
}
\description{
A function that executes the whole workflow of creating/updating gh-pages content from a data model.
}
