% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_contexts.R
\name{main_contexts}
\alias{main_contexts}
\title{Main}
\usage{
main_contexts(portal, template_dir = "model_templates", branch = "./")
}
\arguments{
\item{portal}{a string indicating the abbreviation of the portal: 'ark', 'veoibd'}

\item{template_dir}{a string specifying the subdir where template files are stored, default = "model_templates"}

\item{branch}{a OPTIONAL string indicating subdir to which the model main branch has been downloaded to, default = "./"}
}
\description{
A function that executes the whole workflow of creating/updating gh-pages content from a data model.
}
