% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{makeTemplateContent}
\alias{makeTemplateContent}
\title{Make metadata collection template content}
\usage{
makeTemplateContent(model, template_dir_path, portal = NULL)
}
\arguments{
\item{model}{a data.frame object containing the data model.}

\item{template_dir_path}{a string specifying the subdir where template files are stored}
}
\description{
A function that executes a series of steps to create/update metadata collection template pages.
}
