% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{makeTemplateContent}
\alias{makeTemplateContent}
\title{Make metadata collection template content}
\usage{
makeTemplateContent(model)
}
\arguments{
\item{model}{a data.frame object containing the data model.}
}
\description{
A function that executes a series of steps to create/update metadata collection template pages.
}
